Dear Contributors
Apologized for rudundance update

this model, however atlast I have

found the point that making

previous model didn't run

completely and result in one

parameter being not process

it's the error under my nose

and currently I have corrected

that point and the parameter

is being process and estimate

as should be.

Then the next step would be

to process the model with

difference interactions and

scenarios that I will keep

doing from now.

If you have any instructions

and/or recommendations

please let me known.

Best Regards
Ghan Saridnirun