Fixed time indexing of NB count data
Added +1 to the time indices of the non-breeding season count data. That is because we build the process model under the assumption that the "peafowl year" starts at the beginning of May = beginning of non-breeding season. Since our data begins in a breeding season (t = 1), the time index of the first non-breeding season is t = 2.